third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-cssom-apis.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
div {
    width: 100px;
    @apply --foo;
    height: 100px;
}
</style>
<script>
var rule = document.styleSheets[0].cssRules[0];
var style = rule.style;

test(function(){
    var expected = "width: 100px; @apply --foo; height: 100px;"
    assert_equals(rule.cssText, "div { " + expected + " }");
    assert_equals(style.cssText, expected);
}, "@apply should serialize correctly in cssText");

test(function(){
    assert_equals(style.item(0), "width");
    assert_equals(style.item(1), "@apply");
    assert_equals(style.item(2), "height");
}, "CSSStyleDeclaration.item should work with @apply");
</script>
